chore: moved datamodel to model package
